Comments resolved -Made seperate component for Task List -Implemented early return when todo list is empty
